Cambios en la vista detallada de las monedas

+++ app/src/main/java/cripto/potencial/tjdian/features/cryptocurrency/data/remote/detailremoteapimodel/MarketCap.kt
@@ -21,7 +21,7 @@ data class MarketCap(
     val dot: Float,
     val eos: Float,
     val eth: Double,
-    val eur: Float,
+    val eur: Long,
     val gbp: Float,
     val hkd: Float,
     val huf: Float,
